+++ src/DuskServiceProvider.php
@@ -2,7 +2,6 @@
 
 namespace Laravel\Dusk;
 
-use Exception;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\ServiceProvider;
 
@@ -15,20 +14,22 @@ class DuskServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        Route::get('/_dusk/login/{userId}/{guard?}', [
-            'middleware' => 'web',
-            'uses' => 'Laravel\Dusk\Http\Controllers\UserController@login',
-        ]);
+        if (! $this->app->environment('production')) {
+            Route::get('/_dusk/login/{userId}/{guard?}', [
+                'middleware' => 'web',
+                'uses' => 'Laravel\Dusk\Http\Controllers\UserController@login',
+            ]);
 
-        Route::get('/_dusk/logout/{guard?}', [
-            'middleware' => 'web',
-            'uses' => 'Laravel\Dusk\Http\Controllers\UserController@logout',
-        ]);
+            Route::get('/_dusk/logout/{guard?}', [
+                'middleware' => 'web',
+                'uses' => 'Laravel\Dusk\Http\Controllers\UserController@logout',
+            ]);
 
-        Route::get('/_dusk/user/{guard?}', [
-            'middleware' => 'web',
-            'uses' => 'Laravel\Dusk\Http\Controllers\UserController@user',
-        ]);
+            Route::get('/_dusk/user/{guard?}', [
+                'middleware' => 'web',
+                'uses' => 'Laravel\Dusk\Http\Controllers\UserController@user',
+            ]);
+        }
     }
 
     /**
@@ -40,10 +41,6 @@ class DuskServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        if ($this->app->environment('production')) {
-            throw new Exception('It is unsafe to run Dusk in production.');
-        }
-
         if ($this->app->runningInConsole()) {
             $this->commands([
                 Console\InstallCommand::class,